+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 0.3.0 :
 
 - Added variable `php_band_php_extension_dir`
+- Added extension building support
 
 0.2.0 :